CDK-445: Allow hive repos to open any Kite table.
This removes the check in load that enforces a managed provider only
opens managed tables, and a similar check in external. The behavior is
tested in TestManagedExternalHandling.

The only change to the repository classes is to move delete into the
abstract repository and make it detect whether the table is external or
managed and use the correct implementation.

This also fixes update, which dropped property updates.