CDK-265: Filter listed Hive tables that are not readable.
This determines whether a table is readable by checking the type and
getting a descriptor for it. That ensures that the table uses supported
input and output formats and that the schema is available.

When converting from a table to a descriptor, HiveUtils now checks
whether the Hive table is partitioned. If so, a partition strategy
expression is requried. This prevents listing tables that can't be read
because they are not partitioned with Kite. In the future, we can
translate from Hive partition columns to a Kite strategy, but because
there is no way to know how the partition columns are derived, they
should be provided partitioners, which is not yet implemented.