CDK-492: Make durable writes default and clean up.
Durable writes should be the default when writing to parquet, and not
using them should be an optimization. This switches the
kite.parquet.durable-writes to kite.parquet.non-durable-writes and adds
the setting to the DatasetKeyOutputFormat because durable writes are not
required when using an OutputCommitter in MR where failed task attempts
are run again.

This also refactors the FS descriptor properties into a common
FileSystemProperties class and moves the isTrue method into a
DescriptorUtil class in the SPI.