CDK-583: Fix non-string types with id partitioner.
The conversion from path to StorageKey was incorrectly using the
FieldPartitioner#getType() method. The fix is to use
SchemaUtil.getPartitionType(), which will use the type from the schema
for identity partitioners rather than Object.class which results in no
conversion.

Most of the changes are to add a Schema to the PathConversion class and
update its tests.