CDK-602: Remove DatasetRepository from kite-data-core.
This removes DatasetRepository from the API along with any dependent
classes, like DatasetRepositoryException and
FileSystemDatasetRepository. The exception has been replaced with more
consistent usage of DatasetIOException and ValidationException, and with
a new exception, DatasetOperationException that wraps any
implementation-specific failure.

RandomAccessDatasetRepository is removed in both the API and SPI. It is
no longer needed when using Datasets methods to instantiate datasets
because the return type is cast based on the context.