CDK-435: Update HBase to use EntityAccessor for keys.
This updates PartitionKey so that its data is extracted using the
EntityAccessor class. It also consolidates key building in the HBase
implementation by adding EntityComposer#extractKey and delegating all
key building to that method (via the new EntityMapper#mapToKey(E)). The
EntityComposer method is based on EntityAccessor because the most
reasonable place to manage the accessor is inside EntityComposer.

AvroEntityComposer also uses the accessor for extractField now, but
checks to ensure that the field is not a nested field, for now. The
getDefaultPrimitive method is no longer used so that this is compatible
with Avro's behavior.