CDK-657: Fix URI qualification in descriptor.
When makeQualified is called on a file URI and the default FS is a HDFS
URI, the authority section from the HDFS URI is incorrectly copied to
the file path. This catches the situation where the default FS scheme
and the location scheme do not match, so the default isn't useful, and
uses the location URI as the default to avoid this bug.