Fix EntityAccessor for read schemas from AllowNull.
The EntityAccessor assumes that the dataset schema has no nullable
record fields for partition source data, which is validated by the
descriptor checks. However, when using reflected records with Schemas
generated by AllowNull, this may not be the case even though the
descriptor has validated correctly. This adds support to skip nullable
unions.