CDK-735: Fix Crunch dataset writes with provided partitioners.
This updates DatasetKeyOutputFormat so that the view constraints are
used when writing to temporary datasets. That ensures that the values
provided by the target view are present so writes to the temporary
datasets don't fail. Fix and test contributed by Ryan Brush.

This also updates the partitioning methods in CrunchDatasets to fix the
CLI copy command when writing to a provided partition. In that case, the
mapper, GetStorageKey, needs the view constraints as well as the
partition strategy in order to create storage keys. The constraints are
serialized using the query map as a stand-in.