CDK-787: Fix Hive backward-compatibility check.
For backward-compatibility with pre-namespaces datasets, Hive dataset
creation checked whether or not a new dataset was replacing an existing
one in the default namespace. If it is, then the schema for the new
dataset is validated to ensure it can read existing data.

The problem was that this check used resolveNamespace, which returns the
default namespace as a match if either the location is null or the
location matches. But the location is almost always null, so this
prevented any table from being created when there is a duplicate name in
the default db. The fix is to consider only new tables that point to a
location matching the existing default table to match for this
validation.