CDK-797: Add CSV field ordering based on header.
This will create records using the field names defined in the CSV
header, if present. It also removes the unused option to turn on this
feature because fields should always be ordered by header if it is
present. To turn off header-based ordering, users can set the CSV
properties to use no header and skip the header line if it is present.