CDK-801: Throw NumberFormatException in CSV reader.
This fixes a bug where the read schema cannot read a numeric value and
null is returned by makeValue. This is surprising behavior because the
read may continue as though nothing happened, but users expect either an
error or for data to be coerced. Instead, the default value would be
used or an unhelpful error message (that there is no default value)
would be shown instead of the underlying error that, for example, float
data can't be read as a long.