CDK-840: Fix flaky CLI tests.
CDK-796 updated Main#run to set Main's Configuration as the
DefaultConfiguration. This is correct for normal use, but caused CLI
tests to be flaky because this global state persists across test cases.

This commit fixes the problem by not calling DefaultConfiguration.set if
the conf from getConf is null (fixing a NPE) and also defaults to using
DefaultConfiguration.get() instead of new Configuration() in the helper
run methods. This ensures that the global Configuration set in
@BeforeClass methods persists for the entire test suite.