CDK-844: Update delete command to delete views.
The delete command will now detect view URIs and delete just the content
using View#deleteAll (subject to deleteAll's limitations). This is
guarded by a check that validates the view matches the requested URI
because this is destructive. If there is an option in the requested URI
that isn't in the view's URI, then the command fails. This catches typos
in field names that are not recognized as real constraints and not added
to the view.