CDK-677: Add translation from Hive DDL to Schema.
Tables created in Hive using STORED AS AVRO do not have a schema URI or
literal in the table properties. Instead, the schema is generated from
the table DDL when writing. This translates from table DDL if no schema
is set. This also fixes tables created with STORED AS PARQUET.

HiveUtils methods for schema conversion have been moved into
HiveSchemaConverter. Tests for Hive tables or types to Avro have been
added as TestTableConversion.