CDK-861: Fix flaky FlumeConfigCommand test.
TestUtil.run calls Main, which would assume that it is actually the main
program and replaces the DefaultConfiguration. This updates run to
replace the default conf with the original after it runs Main. Two tests
used TestUtil.run and passed a new Configuration, which caused the tests
that depended on the default FS URI setting from @BeforeClass to fail.
These do not seem to depend on the new configuration, so I've removed
it.