CDK-188: Pass descriptor properties to InputFormat.
This update the MR code so that the InputFormat is passed the descriptor
properties of the dataset descriptor via the job Configuration. This
allows users to configure the InputFormat used by Kite by setting
descriptor properties. This is needed to wrap an InputFormat dataset in
Kite and be able to read it.

This also update the InputFormat configuration so that the InputFormat
class is always added to the Configuration, even if
configure(Configuration) is used instead of configure(Job).