CDK-188: Fix hadoop-1 InputFormat setting.
The last commit updated DatasetKeyInputFormat.configure(Configuration)
to correctly set the InputFormat by creating a fake Job and copying the
configuration settings produced by calling Job.setInputFormatClass. In
Hadoop 1, this breaks Crunch because there are extra properties set.

The behavior to get the properties set by setInputFormatClass are
correct in the other cases because the property is in MRJobConfig and is
marked as private/evolving. In other cases, only one property is set.