CDK-885: Fix FS writer error handling.
This updates the FileSystemWriter's error handling. First,
RuntimeException is now caught throughout the class to ensure that
unknown exceptions in appenders (or FS calls) cannot put the writer in
an unknown state. Second, this updates the close method so that it
handles more error cases correctly:

* After a call to close, the writer state is always CLOSED
* Close does nothing when writer is NEW or CLOSED
* Close is called (on Appender) only if the writer is not in ERROR
* A data file is released if:
  1. Data has been written to the file
  2. Data has been committed with a flush() or sync()
  3. The writer is not in an ERROR state
* If a data file is not to be released, it is discarded
* Appender cleanup is called for OPEN or ERROR
* Any failure in close will prevent the next close task
* No close tasks will be retried b/c state is CLOSED

This also updates the TestFileSystemWriters tests and adds tests for
Parquet, which were previously missing.