CDK-766: Add CSV header CLI option.
This adds --header to csv-schema and csv-import. The option is a single
string that is passed to the CSV code in kite-data-core in
CSVProperties. The header can be stored as a new descriptor property,
kite.csv.header.

The CSVProperties header is always low-precedence. If useHeader is set
in properties, a header line is read from the file or data sample (for
inferSchema) and used instead. For CSVRecordParser, the properites
header is only used if the header passed is null.