CDK-955: Fix duplicate URIs from Datasets.list.
The problem was that the FileSystem implementation was returning
multiple "default" namespaces. The implementation now uses a Set to
accumulate both namespaces and dataset names.

The bug only happened when a repository namespace path was used instead
of a repository. In that case, the single namespace appears to be a
valid repository with tables in the default namespace. Each table in the
default namespace would result in another "default" in the list of
namespaces. When Datasets.list listed the tables in each namespace, all
of the tables were present in each copy of default, which resulted in
n^2 results.