CDK-898: Fix file not found bug in distributed cache.
Setup in the distributed cache is failing when using the local job
runner when HDFS is the default file system. This is very likely related
to HDFS-7031. The work-around is to set the local FS as the default when
using LocalJobRunner. This is safe because HDFS has already been loaded,
so lookups without the authority with still resolve to the full URI.