KITE-991: Add SPI calls to set roll configuration.
This adds RollingWriter to the SPI, which adds setRollIntervalMillis and
setTargetFileSize methods. The partitioned writer and file writer now
implement this to allow callers to check whether the writer is a rolling
writer and configure rolling without setting descriptor properties.

The descriptor properties are used to initialize the roll interval and
target file size, but are overridden by the setter methods.